refactor: move constants/error utils to common
